+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsApi.kt
@@ -1,5 +1,9 @@
 package com.braintreepayments.api.core
 
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
@@ -12,6 +16,8 @@ internal class AnalyticsApi(
     private val deviceInspector: DeviceInspector = DeviceInspectorProvider().deviceInspector,
     private val analyticsParamRepository: AnalyticsParamRepository = AnalyticsParamRepository.instance,
     private val merchantRepository: MerchantRepository = MerchantRepository.instance,
+    private val dispatcher: CoroutineDispatcher = Dispatchers.Main,
+    private val coroutineScope: CoroutineScope = CoroutineScope(dispatcher)
 ) {
 
     fun execute(
@@ -25,14 +31,16 @@ internal class AnalyticsApi(
             sessionId = analyticsParamRepository.sessionId,
             integration = merchantRepository.integrationType
         )
-        val analyticsRequest = createFPTIPayload(merchantRepository.authorization, jsonEvents, metadata)
-        httpClient.post(
-            path = FPTI_ANALYTICS_URL,
-            data = analyticsRequest.toString(),
-            configuration = null,
-            authorization = merchantRepository.authorization,
-            callback = null
-        )
+        val analyticsRequest =
+            createFPTIPayload(merchantRepository.authorization, jsonEvents, metadata)
+        coroutineScope.launch {
+            httpClient.post(
+                path = FPTI_ANALYTICS_URL,
+                data = analyticsRequest.toString(),
+                configuration = null,
+                authorization = merchantRepository.authorization,
+            )
+        }
     }
 
     @Throws(JSONException::class)
@@ -89,11 +97,13 @@ internal class AnalyticsApi(
             .putOpt(FPTI_KEY_ERROR_DESC, event.errorDescription)
             .putOpt(FPTI_KEY_CONTEXT_TYPE, if (event.isVaultRequest) "BA-TOKEN" else "EC-TOKEN")
             .putOpt(FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH, event.didSdkAttemptAppSwitch)
+            .putOpt(FPTI_KEY_FUNDING_SOURCE, event.fundingSource)
     }
 
     @Throws(JSONException::class)
     private fun mapDeviceMetadataToFPTIBatchParamsJSON(metadata: DeviceMetadata): JSONObject {
-        val isVenmoInstalled = deviceInspector.isVenmoInstalled(merchantRepository.applicationContext)
+        val isVenmoInstalled =
+            deviceInspector.isVenmoInstalled(merchantRepository.applicationContext)
         val isPayPalInstalled = deviceInspector.isPayPalInstalled()
         return metadata.run {
             JSONObject()
@@ -146,6 +156,7 @@ internal class AnalyticsApi(
         private const val FPTI_KEY_PAYPAL_ATTEMPTED_APP_SWITCH = "attempted_app_switch"
         private const val FPTI_KEY_ERROR_DESC = "error_desc"
         private const val FPTI_KEY_CONTEXT_TYPE = "context_type"
+        private const val FPTI_KEY_FUNDING_SOURCE = "funding_source"
 
         private const val FPTI_BATCH_KEY_VENMO_INSTALLED = "venmo_installed"
         private const val FPTI_BATCH_KEY_PAYPAL_INSTALLED = "paypal_installed"