+++ BraintreeCore/src/test/java/com/braintreepayments/api/core/AnalyticsParamRepositoryUnitTest.kt
@@ -14,6 +14,7 @@ class AnalyticsParamRepositoryUnitTest {
 
     private val uuid = "test-uuid"
     private val newUuid = "new-uuid"
+    private val expectedFundingSource = "funding-source"
 
     @Before
     fun setUp() {
@@ -24,6 +25,7 @@ class AnalyticsParamRepositoryUnitTest {
         sut.didPayPalServerAttemptAppSwitch = true
         sut.didSdkAttemptAppSwitch = true
         sut.didEnablePayPalAppSwitch = true
+        sut.fundingSource = expectedFundingSource
     }
 
     @Test
@@ -43,6 +45,7 @@ class AnalyticsParamRepositoryUnitTest {
         assertEquals(true, sut.didPayPalServerAttemptAppSwitch)
         assertEquals(true, sut.didEnablePayPalAppSwitch)
         assertEquals(true, sut.didSdkAttemptAppSwitch)
+        assertEquals(expectedFundingSource, sut.fundingSource)
 
         sut.reset()
 
@@ -50,5 +53,6 @@ class AnalyticsParamRepositoryUnitTest {
         assertNull(sut.didPayPalServerAttemptAppSwitch)
         assertNull(sut.didEnablePayPalAppSwitch)
         assertNull(sut.didSdkAttemptAppSwitch)
+        assertNull(sut.fundingSource)
     }
 }